Add tests for MinMax with selection overloads (generic)
